networkd: make Dhclient.stop to stop the DHCP client managing the interface if running and to unconfigure addresses.
No functional changes expected (only `Ip.flush_ip_addr` could be called twice instead of once).

+++ ocaml/networkd/lib/network_utils.ml
@@ -918,6 +918,7 @@ module Dhclient : sig
   val is_running : ?ipv6:bool -> interface -> bool
 
   val stop : ?ipv6:bool -> interface -> unit
+  (** stop: stop the DHCP client managing [interface] if running and to unconfigure addresses. *)
 
   val ensure_running :
        ?ipv6:bool
@@ -1050,27 +1051,33 @@ end = struct
         ]
       )
 
-  let stop ?(ipv6 = false) interface =
-    try
-      ignore
-        (call_script dhclient
-           [
-             "-r"
-           ; "-pf"
-           ; pid_file ~ipv6 interface
-           ; "-lf"
-           ; lease_file ~ipv6 interface
-           ; interface
-           ]
-        ) ;
-      Unix.unlink (pid_file ~ipv6 interface)
-    with _ -> ()
-
   let is_running ?(ipv6 = false) interface =
     try
       Unix.access (pid_file ~ipv6 interface) [Unix.F_OK] ;
       true
     with Unix.Unix_error _ -> false
+
+  let stop ?(ipv6 = false) interface =
+    if is_running ~ipv6 interface then (
+      Xapi_stdext_pervasives.Pervasiveext.ignore_exn (fun () ->
+          (* release DHCP lease and close the DHCP client *)
+          ignore
+            (call_script dhclient
+               [
+                 "-r"
+               ; "-pf"
+               ; pid_file ~ipv6 interface
+               ; "-lf"
+               ; lease_file ~ipv6 interface
+               ; interface
+               ]
+            ) ;
+          (* remove the pid file *)
+          Unix.unlink (pid_file ~ipv6 interface)
+      ) ;
+      (* flush configured addresses *)
+      Ip.flush_ip_addr ~ipv6 interface
+    )
 
   let ensure_running ?(ipv6 = false) interface options =
     if not (is_running ~ipv6 interface) then
@@ -1082,7 +1089,7 @@ end = struct
       let current_conf = read_conf_file ~ipv6 interface in
       let new_conf = generate_conf ~ipv6 interface options in
       if current_conf <> Some new_conf then (
-        ignore (stop ~ipv6 interface) ;
+        stop ~ipv6 interface ;
         ignore (start ~ipv6 interface options)
       )
 end
